+++ src/actions/FirstOpenActions.tsx
@@ -1,17 +1,15 @@
-import { asNumber, asObject, asOptional, asString, asValue } from 'cleaners'
+import { asNumber, asObject, asString, asValue } from 'cleaners'
 import { makeReactNativeDisklet } from 'disklet'
 
 import { FIRST_OPEN } from '../constants/constantSettings'
 import { makeUuid } from '../util/rnUtils'
-import { getCountryCodeByIp } from './AccountReferralActions'
 
 const firstOpenDisklet = makeReactNativeDisklet()
 
 const asFirstOpenInfo = asObject({
   isFirstOpen: asValue('true', 'false'),
   deviceId: asString,
-  firstOpenEpoch: asNumber,
-  countryCode: asOptional(asString)
+  firstOpenEpoch: asNumber
 })
 type FirstOpenInfo = ReturnType<typeof asFirstOpenInfo>
 
@@ -42,15 +40,9 @@ const readFirstOpenInfoFromDisk = async (): Promise<FirstOpenInfo> => {
       firstOpenText = await firstOpenDisklet.getText(FIRST_OPEN)
       firstOpenInfo = asFirstOpenInfo(JSON.parse(firstOpenText))
       firstOpenInfo.isFirstOpen = 'false'
-
-      if (firstOpenInfo.countryCode == null) {
-        // Not critical if we can't get the country code
-        firstOpenInfo.countryCode = await getCountryCodeByIp().catch(() => undefined)
-      }
     } catch (error: unknown) {
       // Generate new values.
       firstOpenInfo = {
-        countryCode: await getCountryCodeByIp(),
         deviceId: await makeUuid(),
         firstOpenEpoch: Date.now(),
         // If firstOpen != null: This is not the first time they opened the app,